Minimum space detection method.

Former-commit-id: c16b7bc1c31a619af5da2d6c3f0adf74ae45fa59 [formerly d9f97b0295633dedc4afaad8171d62ea9e576401]
Former-commit-id: 46a0e08ae63042f9059bf0bdbc350573f1afc140

+++ blockdev.cpp
@@ -39,7 +39,7 @@ void BlockDeviceInfo::addToCombo(QComboBox *combo, bool warnNasty) const
 {
     const char *suffixes[] = {"B", "KB", "MB", "GB", "TB", "PB", "EB"};
     unsigned int isuffix = 0;
-    qlonglong scalesize = size;
+    long long scalesize = size;
     while (scalesize >= 1024 && isuffix < sizeof(suffixes)) {
         ++isuffix;
         scalesize /= 1024;
